now all times are in ms and uint

+++ NaoTHSoccer/Source/Cognition/Modules/Modeling/BallLocator/TeamBallLocator/TeamBallLocator.h
@@ -50,7 +50,7 @@ class TeamBallLocator : protected TeamBallLocatorBase
         }
       public:
         Vector2d vec;
-        double t;
+        unsigned int t;
     };
 
   public:
